# handlers/apidocs/openapi.yaml
openapi: 3.0.0
info:
  title: Pi API
  description: The endpoints go-server took over from the Raspberry Pi's old Node webserver.
  version: 1.0.0
servers:
  - url: http://raspberrypi.local:8080
    description: Raspberry Pi
paths:
...
  /monitor/history:
    get:
      summary: Readings over a time range, in columns
      description: >
        Up to 15 minutes comes from the 5-second readings in memory (tier
        "live"). Longer ranges come from the finest stored tier that covers
        them: per-minute readings (48 hours), 5-minute rollups (30 days) or
        hourly rollups (2 years). Rollups also carry each metric's peak as
        "<metric>_max". Unknown values are null.
      parameters:
        - name: range
          in: query
          required: false
          description: A number and a unit (m, h, d or y), at most 2y.
          schema: { type: string, default: 1h, example: 7d }
      responses:
        '200':
          description: The readings
          content:
            application/json:
              schema:
                type: object
                properties:
                  tier: { type: string, enum: [live, samples, rollup_5m, rollup_1h] }
                  step: { type: integer, description: Seconds between readings }
                  ts: { type: array, items: { type: integer }, description: Unix seconds }
                  series:
                    type: object
                    additionalProperties:
                      type: array
                      items: { type: number, nullable: true }
        '400':
          description: The range could not be read
  /monitor/events:
    get:
      summary: Service status changes, newest first
      description: A change is recorded once the new status has held for two checks in a row. Kept 90 days.
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 50, maximum: 500 }
      responses:
        '200':
          description: The changes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    ts: { type: integer }
                    group: { type: string }
                    name: { type: string }
                    status: { type: string, enum: [up, warn, down, stopped] }
                    detail: { type: string }
...